refactor(constants): migrate chaos and utilities imports to submodules
Replace `from utilities.constants import X` with direct submodule imports
in tests/chaos/ and utilities/*.py. Images stays on the package shim
until the compat shim is dropped.

Generated-by: Claude Sonnet 4.6
