+++ tests/chaos/standard/test_standard.py
@@ -3,13 +3,13 @@ from ocp_resources.deployment import Deployment
 from ocp_resources.virtual_machine import VirtualMachine
 
 from tests.chaos.constants import STRESS_NG
-from utilities.constants import (
+from utilities.constants import Images
+from utilities.constants.namespaces import NamespacesNames
+from utilities.constants.storage import StorageClassNames
+from utilities.constants.timeouts import (
     TIMEOUT_2MIN,
     TIMEOUT_5MIN,
     TIMEOUT_5SEC,
-    Images,
-    NamespacesNames,
-    StorageClassNames,
 )
 from utilities.virt import VirtualMachineForTests, running_vm
