+++ utilities/artifactory.py
@@ -9,7 +9,10 @@ from ocp_resources.secret import Secret
 from pytest_testconfig import config as py_config
 from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
-from utilities.constants import TIMEOUT_1MIN, TIMEOUT_5SEC
+from utilities.constants.timeouts import (
+    TIMEOUT_1MIN,
+    TIMEOUT_5SEC,
+)
 from utilities.data_utils import base64_encode_str
 
 LOGGER = logging.getLogger(__name__)
